chore!: remove the Map trait from public API
