+++ src/read_only.rs
@@ -1,5 +1,4 @@
 use crate::lock::RwLock;
-use crate::t::Map;
 use crate::{DashMap, HashMap};
 use cfg_if::cfg_if;
 use core::borrow::Borrow;
