+++ src/set.rs
@@ -265,7 +265,7 @@ impl<'a, K: 'a + Eq + Hash, S: BuildHasher + Clone> DashSet<K, S> {
     /// words.insert("hello");
     /// assert_eq!(words.iter().count(), 1);
     /// ```
-    pub fn iter(&'a self) -> Iter<'a, K, S, DashMap<K, (), S>> {
+    pub fn iter(&'a self) -> Iter<'a, K, S> {
         let iter = self.inner.iter();
 
         Iter::new(iter)
